CPP09ex00 better

+++ CPP09/ex00/BitcoinExchange.hpp
@@ -12,5 +12,5 @@
 
 class BitcoinExchange {
 	public:
-		std::map<time_t, int> _data;
+		std::map<time_t, double> _data;
 };
